+++ CHANGELOG.md
@@ -14,7 +14,8 @@ All notable changes to this project will be documented in this file.
 - Stats: top albums
 - Stats: open full top listings
 - Video: option for muted auto play
-- Slideshow: option for no transition
+- Slideshow / Screen saver: option for no transition
+- Slideshow / Screen saver: animated zoom effect
 - Widget: tap action setting
 - Wallpaper: scroll effect option